security: add IPE prototype to kernel tree
IPE is a Linux Security Module, which allows configurable enforcement of
rules that validate code and data integrity on a system.

Integrity requirements can be specified as rules on a set of properties.
A property can be the signature on the artifact (attached/detached
signature on binary), volume property (read-only etc.), device mapper
property (roothash of dm-verity tree must be signed and verified  etc.) or
some other mechanism. The LSM architecture facilitates easy addition of new
properties.

This version of IPE is a prototype and should be backed out when the
upstream version is available.

This version of IPE contains the following erratas:

1. add errata around anonymous memory

>  Foreign Function Interface libraries are causing issues by creating
>  dynamically generated code in anonymous segments. Temporarily allow
>  anonymous mapped-code to pass IPE, but audit the failure as if it was
>  disallowed.

2. ipe: add overlake errata around kexec

>  IPE no longer monitors KEXEC_IMAGE and KEXEC_INITRAMFS on it's
>  kernel_read hook. It assumes that IMA will take care of the
>  verification.

Signed-off-by: Deven Bowers <deven.desai@linux.microsoft.com>
Signed-off-by: Pavel Tatashin <pasha.tatashin@soleen.com>