Add Block Device Flag for DMVerity
Adds the block device flag BD_DM_VERITY_VERIFIED_DEV that
is set when a block device's root hash is verified by dmverity during
the device-mapper block device creation process.

The intended use case of this flag is for integrity subsystems to
identify whterh the origin of a file is from a block device that is
dmverity verified.

Signed-off-by: Deven Bowers <deven.desai@linux.microsoft.com>
Signed-off-by: Pavel Tatashin <pasha.tatashin@soleen.com>